volume info headers; drive size
-backwards-compatible volume info headers
- ulong? drive size from CIM

+++ SearchDirLists/DoubleFile/DoubleFile/SaveDirListings/SaveDirListing.cs
@@ -28,7 +28,7 @@ namespace DoubleFile
 
                 var strModel = m_volStrings.DriveModel;
                 var strSerialNo = m_volStrings.DriveSerial;
-                int? nSize = null;
+                ulong? nSize = null;
 
                 if (string.IsNullOrWhiteSpace(strModel))
                 {
